Tests for workday browser controls
